feat: Adding types around references
