+++ packages/cozy-client/src/index.js
@@ -41,10 +41,3 @@ export * from './mock'
 
 export * from './hooks'
 export { BulkEditError } from './errors'
-
-export {
-  isReferencedBy,
-  isReferencedById,
-  getReferencedBy,
-  getReferencedById
-} from './queries/referencedBy'
