+++ packages/cozy-client/types/associations/index.d.ts
@@ -6,4 +6,4 @@ import HasOneInPlace from "./HasOneInPlace";
 import HasManyInPlace from "./HasManyInPlace";
 import HasManyTriggers from "./HasManyTriggers";
 export { Association, HasManyFiles, HasMany, HasOne, HasOneInPlace, HasManyInPlace, HasManyTriggers };
-export { resolveClass, create } from "./helpers";
+export { resolveClass, create, isReferencedBy, isReferencedById, getReferencedBy, getReferencedById } from "./helpers";
